+++ frontend/src/components/Categories/Categories.css
@@ -1,0 +1,69 @@
+.categories {
+    margin-top: 15px;
+    display: flex;
+    padding: 0 20px;
+}
+
+.category {
+    height: 60px;
+    display: flex;
+    padding: 0 25px;
+    align-items: center;
+    background: var(--direct-category);
+    font-weight: 700;
+    font-size: 16px;
+    background-clip: padding-box;
+}
+
+.category-button {
+    position: relative;
+    height: 50px;
+    width: 50px;
+    border-radius: 15px 15px 0 0;
+    z-index: 2;
+
+
+    &.active {
+        background: var(--direct-category);
+        background-size: 650%;
+        background-position: -25px;
+
+        .radius {
+            position: absolute;
+            left: -25px;
+            bottom: 0;
+            height: 25px;
+            width: 100px;
+            z-index: 0;
+            pointer-events: none;
+            background: var(--direct-category);
+            background-size: 325%;
+        }
+
+        &::before {
+            content: "";
+            position: absolute;
+            top: 0;
+            left: -50px;
+            right: 50px;
+            bottom: 0;
+            border-radius: 15px;
+            background: var(--background-color);
+            pointer-events: none;
+            z-index: 1;
+        }
+
+        &::after {
+            content: "";
+            position: absolute;
+            top: 0;
+            left: 50px;
+            right: -50px;
+            bottom: 0;
+            border-radius: 15px;
+            background: var(--background-color);
+            pointer-events: none;
+            z-index: 1;
+        }
+    }
+}
