+++ frontend/src/components/ChatList/ChatList.css
@@ -1,0 +1,114 @@
+.chat-list {
+    display: flex;
+    flex-direction: column;
+
+    .chat-button {
+        height: 64px;
+        color: var(--white-text);
+        display: flex;
+        align-items: center;
+        gap: 18px;
+        user-select: none;
+        cursor: pointer;
+        padding: 9px;
+    
+        .avatar-block {
+            background: url(/images/user.png) center no-repeat;
+            background-size: 100%;
+            height: 46px;
+            width: 46px;
+            border-radius: 25px;
+        }
+    
+        &:hover {
+            background-color: var(--light-gray-selection);
+        }
+    }
+    
+    .chat-button.active {
+        background-color: var(--light-gray-selection);
+    }
+}
+
+.create-chat-button {
+    background-color: #252525;
+    height: 50px;
+    width: 100%;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: 0.2s;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    &:hover {
+        background-color: #2e2e2e;
+    }
+}
+
+.create-chat-modal {
+    position: fixed;
+    height: 100%;
+    width: 100%;
+    background: rgba(0, 0, 0, 0.5);
+    display: none;
+    align-items: center;
+    justify-content: center;
+    z-index: 1;
+
+    .modal-content {
+        position: relative;
+        background-color: #1D1D1D;
+        height: 500px;
+        width: 500px;
+        border-radius: 5px;
+        display: flex;
+        flex-direction: column;
+        gap: 20px;
+        padding: 50px;
+    }
+
+    .close-modal {
+        position: absolute;
+        right: 5px;
+        top: 5px;
+        height: 30px;
+        width: 30px;
+        background-color: rgb(43, 43, 43);
+        border-radius: 5px;
+        cursor: pointer;
+
+        &:hover {
+            background-color: rgb(53, 53, 53);
+        }
+    }
+
+    input[type="text"] {
+        height: 25px;
+        background: transparent;
+        border: none;
+        border-bottom: 1px solid rgb(59, 59, 59);
+        font-size: 14px;
+        color: white;
+        padding: 0 8px;
+    }
+
+    input[type="submit"] {
+        height: 40px;
+        background-color: #141414;
+        border: none;
+        border-radius: 5px;
+        padding: 0 20px;
+        position: absolute;
+        bottom: 5px;
+        left: 5px;
+        right: 5px;
+        color: #adadad;
+        cursor: pointer;
+
+        &:hover {
+            background-color: #0f0f0f;
+            color: white;
+        }
+    }
+}
