+++ frontend/src/components/Search/Search.css
@@ -1,0 +1,19 @@
+.search {
+    width: 100%;
+    padding: 10px 20px;
+
+    input {
+        width: 100%;
+        height: 40px;
+        padding: 0 12px;
+        border: 0;
+        background-color: transparent;
+        border-bottom: 1px solid var(--dark-gray-text);
+        color: var(--white-text);
+        font-size: 14px;
+        
+        &::placeholder {
+            color: var(--dark-gray-text);
+        }
+    }
+}
