+++ frontend/src/components/Search/Search.jsx
@@ -1,0 +1,9 @@
+import './Search.css';
+
+export default function Search() {
+    return (
+        <div className='search'>
+            <input type="text" placeholder='search'/>
+        </div>
+    )
+}
