+++ frontend/src/pages/MainPage.css
@@ -1,128 +1,9 @@
-.chats-container {
+.container {
     height: 100vh;
     display: flex;
-    justify-content: center;
-    padding: 10px 0 10px 0;
+    padding: 10px 10px 10px 0;
 }
 
-.chat-list {
-    display: flex;
-    flex-direction: column;
-    padding: 10px;
-    gap: 10px;
-    width: 250px;
-}
-
-.chat-section {
-    display: flex;
-    gap: 10px;
-}
-
-.create-chat-button {
-    background-color: #1D1D1D;
-    height: 50px;
-    width: 100%;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: 0.2s;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &:hover {
-        background-color: #242424;
-    }
-}
-
-.create-chat-modal {
-    position: fixed;
-    height: 100%;
-    width: 100%;
-    background: rgba(0, 0, 0, 0.5);
-    display: none;
-    align-items: center;
-    justify-content: center;
-    z-index: 1;
-
-    .modal-content {
-        position: relative;
-        background-color: #1D1D1D;
-        height: 500px;
-        width: 500px;
-        border-radius: 5px;
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        padding: 50px;
-    }
-
-    .close-modal {
-        position: absolute;
-        right: 5px;
-        top: 5px;
-        height: 30px;
-        width: 30px;
-        background-color: rgb(43, 43, 43);
-        border-radius: 5px;
-        cursor: pointer;
-
-        &:hover {
-            background-color: rgb(53, 53, 53);
-        }
-    }
-
-    input[type="text"] {
-        height: 25px;
-        background: transparent;
-        border: none;
-        border-bottom: 1px solid rgb(59, 59, 59);
-        font-size: 14px;
-        color: white;
-        padding: 0 8px;
-    }
-
-    input[type="submit"] {
-        height: 40px;
-        background-color: #141414;
-        border: none;
-        border-radius: 5px;
-        padding: 0 20px;
-        position: absolute;
-        bottom: 5px;
-        left: 5px;
-        right: 5px;
-        color: #adadad;
-        cursor: pointer;
-
-        &:hover {
-            background-color: #0f0f0f;
-            color: white;
-        }
-    }
-}
-
-.chat-button {
-    height: 50px;
-    color: white;
-    border-radius: 5px;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    user-select: none;
-    cursor: pointer;
-
-    .avatar-block {
-        background: url(/images/avatar.svg) center no-repeat;
-        background-size: 80%;
-        height: 50px;
-        width: 50px;
-    }
-
-    &:hover {
-        background-color: #1D1D1D;
-    }
-}
-
-.chat-button.active {
-    background-color: #1D1D1D;
+.left-panel {
+    min-width: 325px;
 }
