+++ frontend/src/proto/msgchat.proto
@@ -15,101 +15,119 @@ service Conversation {
 
     rpc GetMessages (GetMessagesRequest) returns (GetMessagesResponse);
     rpc SendMessage (SendMessageRequest) returns (SendMessageResponse);
-}
 
-// CreateChat
-message CreateChatRequest {
+    // Bidirectional streaming (WIP)
+    rpc ChatStream (stream ChatStreamRequest) returns (stream ChatStreamResponse);
+  }
+
+  // CreateChat
+  message CreateChatRequest {
     string type = 1;
     string name = 2;
     repeated string user_ids = 3;
-}
-
-message CreateChatResponse {
+  }
+  
+  message CreateChatResponse {
     string chat_id = 1;
-}
-
-// GetUserChats
-message GetUserChatsRequest {
+  }
+  
+  // GetUserChats
+  message GetUserChatsRequest {
     string type = 1;
-}
+  }
 
-message GetUserChatsResponse {
+  message GetUserChatsResponse {
     repeated ChatPreview chats = 1;
-}
+  }
 
 
-// GetChatInfo
-message GetChatInfoRequest {
+  // GetChatInfo
+  message GetChatInfoRequest {
     string chat_id = 1;
-}
+  }
 
-message GetChatInfoResponse {
+  message GetChatInfoResponse {
     string chat_id = 1;
     string type = 2;
     string name = 3; 
     repeated string member_ids = 4;
     repeated Channel channels = 5;
-}
+  }
 
-// CreateChannel
-message CreateChannelRequest {
+  // CreateChannel
+  message CreateChannelRequest {
     string chat_id = 1;
     string name = 2;
     string type = 3;
-}
+    
+  }
 
-message CreateChannelResponse {
+  message CreateChannelResponse {
     string channel_id = 1;
-}
+  }
 
 
-// GetMessages и SendMessage
-message GetMessagesRequest {
+  // GetMessages и SendMessage
+  message GetMessagesRequest {
     string channel_id = 1;
     int32 limit = 2;
     int32 offset = 3;
-}
+  }
 
-message GetMessagesResponse {
+  message GetMessagesResponse {
     repeated Message messages = 1;
-}
+  }
 
-message SendMessageRequest {
+  message SendMessageRequest {
     string channel_id = 1;
     string text = 2;
-}
+  }
 
-message SendMessageResponse {
+  message SendMessageResponse {
     string message_id = 1;
-}
+  }
 
-// Сущности
-message Chat {
+  // ChatStream
+  message ChatStreamRequest {
+    oneof payload {
+      SendMessageRequest send_message = 1;
+      string channel_id = 2;
+    }
+  }
+  
+  message ChatStreamResponse {
+    oneof payload {
+      Message new_message = 1;
+      string error_message = 2;
+    }
+  }
+  // Сущности
+  message Chat {
     string chat_id = 1;
     string type = 2;
     string name = 3;
     repeated string member_ids = 4;
     repeated string channel_ids = 5;
-}
+  }
 
-message ChatPreview {
+  message ChatPreview {
     string chat_id = 1;
     string name = 2;
-}
+  }
 
-message Channel {
+  message Channel {
     string channel_id = 1;
     string chat_id = 2;
     string name = 3;
     string type = 4;
     repeated string message_ids = 5;
-  
-}
+    
+  }
 
-message Message {
+  message Message {
     string message_id = 1;
     string channel_id = 2;
     string text = 3;
     string sender_id = 4;
     google.protobuf.Timestamp created_at = 5; 
-}
+  }
